backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
twikit>=2.3.0
google-genai>=0.2.0
python-dotenv>=1.0.0
